feat: Add --force flag and fix ROUND() casting in ML refresh scripts
## Changes Made

### Refresh Scripts Enhanced
- Added `--force` parameter to both refresh scripts to force model retraining
- Added command-line argument parsing for better UX
- Improved logic for conditional model retraining (Sunday schedule + accidents)
- Added robust table/column existence checks before querying

### SQL Fixes
- Fixed ROUND() function calls to cast DOUBLE PRECISION to NUMERIC
  - `ROUND(AVG(g_force)::NUMERIC, 4)`
  - `ROUND(STDDEV(speed_mph)::NUMERIC, 2)`
  - `ROUND(AVG(speed_mph)::NUMERIC, 2)`
  - `ROUND(MAX(speed_mph)::NUMERIC, 2)`
- Uncommented Step 3 model training in recalculate_safe_driver_scores.sql

### New Files
- `refresh_safe_driver_scores_fixed.sh` - Workaround for corrupted Parquet files
- `fix_corrupted_telemetry.sql` - Creates working external table excluding bad partitions
- `REFRESH_SCRIPT_USAGE.md` - Complete usage documentation

## Benefits
- Force model retraining on-demand with `./refresh_safe_driver_scores.sh --force`
- Handles first-time setup gracefully
- Fixes Greenplum ROUND() type errors
- Provides workaround for corrupted HDFS Parquet files
- Complete end-to-end ML pipeline automation

+++ recalculate_safe_driver_scores.sql
@@ -25,7 +25,7 @@ SELECT
     ) as speed_compliance_rate,
     
     -- Driving smoothness (secondary ML feature - 25% impact)
-    ROUND(AVG(g_force), 4) as avg_g_force,
+    ROUND(AVG(g_force)::NUMERIC, 4) as avg_g_force,
     COUNT(*) FILTER (WHERE g_force > 1.5) as harsh_driving_events,
     
     -- Distraction indicators (15% impact)
@@ -35,11 +35,11 @@ SELECT
     ) as phone_usage_rate,
     
     -- Speed consistency (5% impact)
-    ROUND(STDDEV(speed_mph), 2) as speed_variance,
-    
+    ROUND(STDDEV(speed_mph)::NUMERIC, 2) as speed_variance,
+
     -- Additional metrics for analysis
-    ROUND(AVG(speed_mph), 2) as avg_speed,
-    ROUND(MAX(speed_mph), 2) as max_speed,
+    ROUND(AVG(speed_mph)::NUMERIC, 2) as avg_speed,
+    ROUND(MAX(speed_mph)::NUMERIC, 2) as max_speed,
     COUNT(*) FILTER (WHERE speed_mph > speed_limit_mph + 10) as excessive_speeding_count,
     
     -- Timestamp
@@ -86,7 +86,6 @@ ALTER TABLE driver_ml_training_data_new RENAME TO driver_ml_training_data;
 -- Normally you would retrain weekly or when significant new patterns emerge
 -- =============================================================================
 
-/*
 -- Drop existing model
 DROP TABLE IF EXISTS driver_accident_model CASCADE;
 DROP TABLE IF EXISTS driver_accident_model_summary CASCADE;
@@ -96,14 +95,13 @@ SELECT madlib.logregr_train(
     'driver_ml_training_data',                    -- Training dataset
     'driver_accident_model',                      -- Output model table
     'has_accident',                               -- Target variable (0/1)
-    'ARRAY[1, speed_compliance_rate, avg_g_force, 
-           harsh_driving_events, phone_usage_rate, 
+    'ARRAY[1, speed_compliance_rate, avg_g_force,
+           harsh_driving_events, phone_usage_rate,
            speed_variance]',                      -- Feature vector with intercept
     NULL,                                         -- No grouping
     20,                                           -- Max iterations
     'irls'                                        -- Iteratively Reweighted Least Squares
 );
-*/
 
 -- Step 4: Calculate New Safety Scores
 -- =============================================================================
